Add tab to add browser/webpage in screenspace renderables panel
