+++ public/locales/en/panel-screenspacerenderable.json
@@ -3,10 +3,17 @@
     "title": "Display name",
     "placeholder": "Slide name"
   },
-  "slide-path-input": {
-    "title": "Path or URL to slide",
+  "image-input": {
+    "tab-title": "Image",
+    "title": "Path or URL to image",
     "placeholder": "Path / URL",
-    "button-label": "Add slide"
+    "button-label": "Add"
+  },
+  "website-input": {
+    "tab-title": "Website",
+    "title": "URL to website",
+    "placeholder": "URL",
+    "button-label": "Add"
   },
   "added-slides": {
     "empty-slides": "No active slides",
